chore: move share to modules
